change timeouts to be longs and not ints + use the provided timeout for a blocking request execution + test case for misbehaving executors that don't respect timeouts

+++ src/main/scala/Asynch.scala
@@ -12,7 +12,7 @@ object `package` {
   type Headers = Vector[Header]
   type Body = Option[String]
 
-  type Timeout = Int
+  type Timeout = Long
 
   type RequestModifier = Request => Request
